+++ lithops/standalone/backends/ibm_vpc/ibm_vpc.py
@@ -91,6 +91,7 @@ class IBMVPCBackend:
             logger.debug(f'Could not find VPC cache data in {self.cache_file}')
         elif 'vpc_id' in self.vpc_data:
             self.vpc_key = self.vpc_data['vpc_id'].split('-')[2]
+            self.vpc_name = self.vpc_data['vpc_name']
 
     def _dump_vpc_data(self):
         """
@@ -367,7 +368,7 @@ class IBMVPCBackend:
                     if 'ubuntu-22' in image['name']:
                         self.config['image_id'] = image['id']
 
-        name = self.config.get('instance_name') or f'lithops-master-{self.vpc_key}'
+        name = self.config.get('master_name') or f'lithops-master-{self.vpc_key}'
         self.master = IBMVPCInstance(name, self.config, self.vpc_cli, public=True)
         self.master.public_ip = self.config['floating_ip']
         self.master.instance_id = self.config['instance_id'] if self.mode == ExecMode.CONSUME.value else None
@@ -389,7 +390,8 @@ class IBMVPCBackend:
 
             ins_id = self.config['instance_id']
             if not self.vpc_data or ins_id != self.vpc_data.get('instance_id'):
-                self.config['instance_name'] = self.vpc_cli.get_instance(ins_id).get_result()['name']
+                name = self.vpc_cli.get_instance(ins_id).get_result()['name']
+                self.config['master_name'] = name
 
             # Create the master VM instance
             self._create_master_instance()
@@ -398,8 +400,8 @@ class IBMVPCBackend:
                 'mode': self.mode,
                 'vpc_data_type': 'provided',
                 'ssh_data_type': 'provided',
-                'instance_id': self.config['instance_id'],
-                'instance_name': self.config['instance_name'],
+                'master_name': self.config['master_name'],
+                'master_id': self.config['instance_id'],
                 'floating_ip': self.config['floating_ip']
             }
 
@@ -423,8 +425,9 @@ class IBMVPCBackend:
                 'mode': self.mode,
                 'vpc_data_type': self.vpc_data_type,
                 'ssh_data_type': self.ssh_data_type,
-                'instance_name': self.master.name,
-                'instance_id': '0af1',
+                'master_name': self.master.name,
+                'master_id': self.vpc_key,
+                'vpc_name': self.vpc_name,
                 'vpc_id': self.config['vpc_id'],
                 'subnet_id': self.config['subnet_id'],
                 'security_group_id': self.config['security_group_id'],
@@ -443,14 +446,14 @@ class IBMVPCBackend:
         """
         Deletes all VM instances in the VPC
         """
-        msg = (f'Deleting all Lithops worker VMs in {self.vpc_name}'
+        msg = (f'Deleting all Lithops worker VMs from {self.vpc_name}'
                if self.vpc_name else 'Deleting all Lithops worker VMs')
         logger.info(msg)
 
         def delete_instance(instance_info):
             ins_name, ins_id = instance_info
             try:
-                logger.info(f'Deleting instance {ins_name}')
+                logger.debug(f'Deleting instance {ins_name}')
                 self.vpc_cli.delete_instance(ins_id)
             except ApiException as err:
                 if err.code == 404:
@@ -482,6 +485,13 @@ class IBMVPCBackend:
                 deleted_instances.update(instances_to_delete)
             else:
                 break
+
+        master_pk = os.path.join(self.cache_dir, f"{self.vpc_data['master_name']}-id_rsa.pub")
+        if os.path.isfile(master_pk):
+            os.remove(master_pk)
+
+        if self.vpc_data['vpc_data_type'] == 'provided':
+            return
 
         # Wait until all instances are deleted
         while get_instances():
@@ -540,17 +550,41 @@ class IBMVPCBackend:
                 else:
                     raise err
 
+    def _delete_vpc(self):
+        """
+        Deletes the VPC
+        """
+        if self.vpc_data['vpc_data_type'] == 'provided':
+            return
+
+        msg = (f'Deleting all Lithops VPC resources from {self.vpc_name}'
+               if self.vpc_name else 'Deleting all Lithops VPC resources')
+        logger.info(msg)
+
+        self._delete_subnet()
+        self._delete_gateway()
+
+        if 'vpc_id' not in self.vpc_data:
+            vpcs_info = self.vpc_cli.list_vpcs().get_result()
+            for vpc in vpcs_info['vpcs']:
+                if vpc['name'] == self.vpc_name:
+                    self.vpc_data['vpc_id'] = vpc['id']
+
+        if 'vpc_id' in self.vpc_data:
+            logger.debug(f'Deleting VPC {self.vpc_data["vpc_name"]}')
+            try:
+                self.vpc_cli.delete_vpc(self.vpc_data['vpc_id'])
+            except ApiException as err:
+                if err.code == 404 or err.code == 400:
+                    logger.debug(err)
+                else:
+                    raise err
+
     def _delete_ssh_key(self):
         """
         Deletes the ssh key
         """
-        if not self.vpc_data:
-            return
-
         if self.vpc_data['ssh_data_type'] == 'provided':
-            return
-
-        if 'ssh_key_filename' not in self.vpc_data:
             return
 
         key_filename = self.vpc_data['ssh_key_filename']
@@ -571,35 +605,6 @@ class IBMVPCBackend:
                 else:
                     raise err
 
-    def _delete_vpc(self):
-        """
-        Deletes the VPC
-        """
-        if not self.vpc_data:
-            return
-
-        if self.vpc_data['vpc_data_type'] == 'provided':
-            return
-
-        self._delete_subnet()
-        self._delete_gateway()
-
-        if 'vpc_id' not in self.vpc_data:
-            vpcs_info = self.vpc_cli.list_vpcs().get_result()
-            for vpc in vpcs_info['vpcs']:
-                if vpc['name'] == self.vpc_name:
-                    self.vpc_data['vpc_id'] = vpc['id']
-
-        if 'vpc_id' in self.vpc_data:
-            logger.debug(f'Deleting VPC {self.vpc_data["vpc_id"]}')
-            try:
-                self.vpc_cli.delete_vpc(self.vpc_data['vpc_id'])
-            except ApiException as err:
-                if err.code == 404 or err.code == 400:
-                    logger.debug(err)
-                else:
-                    raise err
-
     def clean(self, all=False):
         """
         Clean all the backend resources
@@ -608,12 +613,19 @@ class IBMVPCBackend:
         logger.info('Cleaning IBM VPC resources')
 
         self._load_vpc_data()
-        self._delete_vm_instances(all=all)
-        self._delete_ssh_key() if all else None
-        self._delete_vpc() if all else None
 
-        if all and os.path.exists(self.cache_file):
-            os.remove(self.cache_file)
+        if not self.vpc_data:
+            return
+
+        if self.vpc_data['mode'] == ExecMode.CONSUME.value:
+            if os.path.exists(self.cache_file):
+                os.remove(self.cache_file)
+        else:
+            self._delete_vm_instances(all=all)
+            self._delete_vpc() if all else None
+            self._delete_ssh_key() if all else None
+            if all and os.path.exists(self.cache_file):
+                os.remove(self.cache_file)
 
     def clear(self, job_keys=None):
         """
@@ -676,7 +688,7 @@ class IBMVPCBackend:
         Creates the runtime key
         """
         name = runtime_name.replace('/', '-').replace(':', '-')
-        runtime_key = os.path.join(self.name, version, self.vpc_data['instance_id'], name)
+        runtime_key = os.path.join(self.name, version, self.vpc_data['master_id'], name)
         return runtime_key
